refactor: rename audit report.ts to display.ts for consistency
Matches the naming convention used by score/display.ts and other
sibling command modules.
