+++ src/shell/commands/audit/display.ts
